Move dialect_shell_complete to sqlfluff.autocomplete
